requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
Werkzeug==3.0.1
email-validator==2.1.0.post1
python-dotenv==1.0.1
Flask-RESTful==0.3.10 